requirements_test.txt
accelerate==1.13.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.12.1
bitsandbytes==0.43.3
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gradio
gradio_client
h11
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.3
# Убираем конкретные версии CUDA библиотек, пусть torch сам ставит нужные
# nvidia-cublas-cu12
# nvidia-cuda-cupti-cu12
# nvidia-cuda-nvrtc-cu12
# nvidia-cuda-runtime-cu12
# nvidia-cudnn-cu12
# nvidia-cufft-cu12
# nvidia-curand-cu12
# nvidia-cusolver-cu12
# nvidia-cusparse-cu12
# nvidia-nccl-cu12
# nvidia-nvjitlink-cu12
# nvidia-nvtx-cu12
opencv-python==4.11.0.86
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic>=2.0.0,<3.0.0
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruff==0.11.5
safetensors==0.5.3
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
starlette==0.46.2
sympy==1.13.1
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.13.2
tqdm==4.67.1
transformers==4.50.3
triton==3.2.0
typer==0.15.3
typing-inspection==0.4.2
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.101
ultralytics-thop==2.0.18
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1